Update multi_file_io.f90
fixed mismatch in integer(4) and integer(8) for lastpos savepos

+++ src/utils_lgpl/deltares_common/packages/deltares_common/src/multi_file_io.f90
@@ -88,7 +88,7 @@
         integer                     :: res
         integer                     :: CUTIL_MF_READ
         integer                     :: strlen, lfindex, crindex, cnullindex
-        integer                     :: lastpos
+        integer(kind=8)             :: lastpos
         character(len=maxRecLength) :: rec
 
         rec = ' '
